refactor(core): route worktree flows through adapters and domain
