+++ src/adapters/git/mod.rs
@@ -2,6 +2,6 @@ pub mod git_worktree_repository;
 pub mod repo_discovery;
 pub mod worktree_lock;
 
-pub use git_worktree_repository::{GitWorktreeManager, WorktreeInfo};
+pub use git_worktree_repository::{CommitInfo, GitWorktreeManager, WorktreeInfo};
 pub use repo_discovery::{open_repository_at_path, open_repository_from_env};
 pub use worktree_lock::WorktreeLock;
